requirements.txt
altgraph==0.17.4
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.1
cycler==0.12.1
DateTime==5.5
dj-database-url==2.2.0
Django
django-cors-headers==4.4.0
django-debug-toolbar==6.0.0
django-filter==24.3
django-otp==1.5.4
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fonttools==4.56.0
gunicorn==22.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.9
pycparser==2.23
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyparsing==3.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
redis==6.4.0
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.6.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.22.0
watchfiles==1.1.0
websockets==15.0.1
zope.interface==7.2
